tcp: Queue tcph to the tap on short packets
Use a TRY..FINALLY block to ensure that tcph gets sent to the
tap even if an exception is raised trying to read the checksum,
urgent pointer, or options. (That mostly means short frames.)

It's not queued earlier because the options can have information
(e.g. SACKs or Multipath TCP) that get added to the struct, but
the listeners can deal with lacking that information, and it's
better than having the taps not receive packets cut short before
the options.

Fix #18138.