epan: do not add display formatted values to the tree
The displayed value for strings obtained using format_text()
should not be confused with the field value being added to the
tree.

proto_tree_add_string() should not be used to add string from
format_text().

First because it is conceptually wrong and secondly because it
will break display filtering with formatted strings with
replacements.

It's also wasteful if it is the only use for the sanitized string
because proto_tree_add_string() will also call format_text() on the
value to fill the display label.