Adding SAP LZC/LZH decompression library
This adds SAP's proprietary implementation of the LZC/LZH compression algorithms, in order
to be used in SAP Diag and RFC dissectors.

The `csdecompr` code is based on [ClamSAP implementation](https://sourceforge.net/p/clamsap/git/ci/main/tree/src/csdecompr.c).
The benefit from using the ClamSAP implementation is that they have already done the work of
porting the original C++ implementation to plain C. In order to mantain the original code as
untouched as possible, the `saplzclzh.c` code acts as a wrapper of the `CsDecompr` "object".
The modifications made to the `csdecompr.c` implementation is only the removal of the
SAR archive file related code. The ClamSAP project is GPLv2 and license headers are kept
intact, with the exception of the wrong contact address that was replaced with the Web site.

The `saplzclzh.c` wrapper is mostly similar to the one present in [SecureAuth's plugin](https://github.com/SecureAuthCorp/SAP-Dissection-plug-in-for-Wireshark/blob/master/src/sapdecompress.cpp).
File `hpa101saptype.h` was kept exactly as in SecureAuth's plugin.

Test code was added using the test cases implemented in [pysap](https://github.com/OWASP/pysap/blob/master/tests/pysapcompress_test.py).