tpm20: make dissector run over fragmented packets correctly
The current tpm20 dissector is assuming a specific fragmentation of
packets (that the command header will be in a 9 byte fragment followed
by the rest of the command in the next fragment, for instance) which
isn't seen in any traces I've taken.  Fix dissection by using the
correct tcp_dissect_pdus() abstraction.

Doing this fix also means we can't distinguish based on fragment size,
so distinguish based on command instead.  Since the command/response
protocol is single threaded, all state can be kept in a single entity
instead of using the wmem btree.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>