tshark: Have dumpcap pipe when run without -w
Add a capture option to make dumpcap write to a data pipe when
capturing. Use that when tshark is capturing and not writing a
capture file (and when no autostop options are set), instead of
writing a temporary file that can grow without bounds.

Create a GIOChannel on the pipe and use that in the wtap reader,
as GIOChannel allows effective and performant polling cross-platform.
Add API calls to wiretap to read with an open GIOChannel. Use
GIOChannels internally for reading based on filenames as well, for
consistency.

This does not yet add the ability for tshark to write file(s) itself
when reading from the data pipe. We would want that eventually for
several cases that cause problems now: applying a read or display
filter and writing out only the files that pass the filter
(#2234, currently not allowed), writing to a pipe or FIFO and
doing dissection (currently not allowed), or writing multiple
output files as a ring buffer and doing dissection (currently can
halt if a file is deleted before tshark reads it, #1650)

Fix #2743.