bHTTP: Introduce Binary HTTP dissector (RFC 9292)
This change introduces a new dissector for RFC 9292 "Binary
Representation of HTTP Messages", bHTTP for short.

bHTTP is a media type of value "message/bhttp". It is a binary format
that bears similarities to HTTP/2 and HTTP/3 frames but is distinct from
them. It uses QUIC variable-length integers.

Detection of bHTTP is based on media type. It can be carried in
conventional protocols, such as HTTP as part of request or response
content. It is also used in Oblivious HTTP, which is left out of scope
of this PR.