CMake: Don't necessarily trust pkg-config about Kerberos
We search through each of the libraries that pkg-config gives
us for Kerberos to find the full paths, because Debian and
Debian derivatives like Ubuntu allowing installing the developer
packages of MIT and Heimdal Kerberos simultaneously.

If we don't actually find the libraries with the hints that
pkg-config gave us, unset all the related variables and don't
claim we have Kerberos.

This is mostly useful when building with USE_STATIC (as happens with
-DOSS_FUZZ=ON), because distributions generally don't ship static
Kerberos libraries.

Fix #18931