epan: Reference count fd_heads in reassembled_table
Sometimes the same key gets reused in a fragment reassembled_table.
In some cases this means we should be using additional key information,
like layer number, though fragment_add_seq_next can trigger this
fairly easily (and it even appears intentional with
reassemble_octet_string in packet-ber.c).

The same reassembled data is entered with multiple keys in the
reassembled table for multiple frames. In order to ensure that
data is deleted when no key refers to it anymore, but also allow
new entries to replace old keys, use reference counting.  It is
simpler than the current approach of freeing all the data at
the end when the table is destroyed, and avoids leaking data.

This is about 95% of the leaks in #19034

(backported from commit 32e175034392f41ce59b97edf517f9635cb42e08)

Reassembly: Attach a replaced tvb to a new one in reassembled tables

Fixup 32e175034392f41ce59b97edf517f9635cb42e08 a bit. We can't
necessarily free a tvb for an entry in the reassembled table that
is being replaced, because there might be an address that directly
shares the tvb memory (thanks to set_address_tvb()).

If we can, add it to the chain for the replacement reassembly's tvb,
so it will still be freed.

Fix #19094

(backported from commit 285a4cb39eb3d066103ba4a8976ca46b2ea41bc2)