epan: Fix setting addr and port on wildcarded conversation
When a conversation lookup not using address B or port B succeeds,
we can try to set the second address or port on the found conversation
if 1) the search lookup *had* an address B or port B, but we weren't
using it for that lookup, and 2) the found conversation *didn't* have
a second address or port.

The second test is done inside conversation_set_[addr,port]2.
The first test is done in find_conversation, but it was accidentally
looking if the *found conversation* had an address or port 2.
This also meant that the port or address setting in those cases
never happened, because first the found conversation was tested
if it had a address or port 2, and then if it didn't yet.