Follow: Improve performance
Use appendPlainText(), which is orders of magnitude faster than
insertPlainText() on large documents. Apparently the latter has
some O(N) algorithm in it to find the current cursor position, even
if at the end of the document. This adds paragraphs instead of text,
so remove the newlines that we add at the end of various types so
that the output data is generally the same (it does add some extra
new lines for ASCII and UTF-8, but that actually makes the output
more similar to that of tshark.)

Create a QByteArray that uses the follow record data as source raw
data, and avoid doing deep copies as much as possible.

Use .at() and .constData() in the C-Array, Hex Dump, and YAML methods
to avoid deep copying the QByteArray. Change the ASCII sanitize_buffer
to take a refefernce to the QByteArray, and use .at() when testing it
so that it only deep copies when a non-ASCII character has to be replaced.

Related to #15963, #17313