epan/dissectors: add unit test support
This adds build system support for unit testing dissectors, using
bt-dht, ip, tcp and udp as examples to show different use cases and to
force initialization of certain features.

All the tests are built into a single executable that is linked agaist
all of libwireshark (through the epan_objects intermediary as some of
the required symbols are not exported in libwireshark) and are run with
the test harness from GLib. By default each test case runs in a
different file and packet scope to try to ensure as much separation as
possible.

The dissector itself is found dynamically with find_(heur_)dissector
avoiding the need to make additional symbols visible, with the downside
that testing at a finer granularity than a whole dissector isn't easily
possible.