WIP:  TLS: Add glob pattern for keylog files
Add glob pattern support in Preferences -> Protocols -> TLS to allow choosing multiple key log files for decryption. This mode is enabled when the last component of the filename contains either '*' or '?'. It will then scan the directory and load any files that match the pattern. Unlike in single log file mode, this mode does not support growing log files. Closes #19069.