capture: Use GIOChannel for stat cache data, use watch
Use a GIOChannel for the stdout/data pipe for stats, and watch
it with g_io_add_watch on the mainloop.
g_io_channel_read_line is suited for reading a line at a time
as the stats use, and this can allow us to get notifications
when the the stats hang up.

This makes it easier to only restart the stats when they die
(reducing the number of dumpcap invocations and hence #15082)
and should make it easier to move all the interface and stats
communication over the stdout pipe (#12222).