matter: track unsecured conversations, report retransmitted packets
Add conversation tracking for unsecured sessions. Conversations are created
based on UDP endpoint pairs as usual, then a list of sessions is stored in
each conversation. For each session, store a mapping between message
counters and Wireshark frame numbers.

This is currently used to implement two user-visible features: adding a
"session index" to ease filtering a single session (similar to tcp.stream),
and marking duplicated message counters as "retransmitted" linking to the
original frame they are a duplicate of.

I'm using arrays and linear search for simplicity. It's unlikely that a
conversation will have so many sessions, or an unsecured session will have
so many messages, that binary search / tree / hashmap will be worth the
overhead. But I may need to revisit this when implementing secure sessions.