Address resolution using SNI found in captured handshake packets
Add a new option (checkbox) to the `Name Resolution` section of the settings
dialog.

When enabling `Use SNI information from captured handshake packets` the Server
Name Indication of TLS connections is used to resolve IPv4 and IPv6 addresses.

This option can be useful when monitoring traffic that uses encrypted DNS.